Add files via upload
I am very glad to share with you guys that I have completed my first task that was assigned to me by CodSoft.
It is a simple Number game where we have to draw random numbers from 0-100 and guess which number it is. 
How will you know what the number is? Well it will simply print a message that whether you have guessed the exact number or you have to guess a higher or a lower number. Then it will tell your total score.
So this task required me to write a program using Java.